refactor: базовый путь переведён на конвейер параметров
PathBaseAppSettings стал чистым POCO, регулярное выражение и его обоснование
переехали в PathBaseAppSettingsValidator.

У UseConfiguredPathBase появился парный AddConfiguredPathBase — единственный
метод пакета, выпадавший из схемы Add*/Use*, исчез. Для потребителей это
ломающее изменение: без вызова Add* метод падает.

Тесты валидации переехали на ValidateOptionsResult, набор входных данных
не изменился.
